requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
billiard==4.1.0
boto3==1.28.20
botocore==1.31.20
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.3
decorator==4.4.2
discord==2.3.1
discord.py==2.3.1
Django==4.1.10
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.1
elevenlabs==0.2.21
executing==1.2.0
flower==2.0.0
frozenlist==1.4.0
future==0.18.3
gcloud==0.18.3
googleapis-common-protos==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
humanize==4.7.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
jmespath==1.0.1
jws==0.1.3
kombu==5.3.1
m3==0.6.12
matplotlib-inline==0.1.6
moviepy==1.0.3
multidict==6.0.4
nose==1.3.7
numpy==1.25.1
oauth2client==4.1.3
openai==0.27.8
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
pymongo==3.12.3
PyMySQL==1.1.0
pyparsing==3.1.1
Pyrebase4==4.7.1
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jwt==2.0.1
pytz==2023.3
redis==4.6.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.2.4
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
yarl==1.9.2
zope.interface==6.1a2
